requirements.txt
PySide6>=6.6.0
numpy>=1.26.0
pyserial>=3.5
